refactor(logic): extract naval mission order helper module (#2178)
Split naval mission-order resolution into a dedicated world module to further decompose naval_resolution.dart while preserving behavior and API surface via re-export.
